test: build e2e/main.go as a binary
