+++ e2e/cluster.go
@@ -1,16 +1,16 @@
-package e2e_test
+package main
 
 import (
 	"context"
 	"fmt"
-	"testing"
+	"log"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armresources"
 )
 
-func setupCluster(ctx context.Context, t *testing.T, cloud *azureClient, location, resourceGroupName, clusterName string) error {
+func setupCluster(ctx context.Context, cloud *azureClient, location, resourceGroupName, clusterName string) error {
 	var needNewCluster bool
 
 	testRgExistence, err := cloud.resourceGroupClient.CheckExistence(ctx, resourceGroupName, nil)
@@ -34,7 +34,7 @@ func setupCluster(ctx context.Context, t *testing.T, cloud *azureClient, locatio
 			}
 
 			if !mcRgExistence.Success || aksCluster.Properties == nil || aksCluster.Properties.ProvisioningState == nil || *aksCluster.Properties.ProvisioningState == "Failed" {
-				t.Log("deleting E2E test cluster in bad state...")
+				log.Println("deleting E2E test cluster in bad state...")
 
 				needNewCluster = true
 				poller, err := cloud.aksClient.BeginDelete(ctx, resourceGroupName, clusterName, nil)
@@ -49,7 +49,7 @@ func setupCluster(ctx context.Context, t *testing.T, cloud *azureClient, locatio
 			}
 		}
 	} else {
-		t.Log("recreating E2E test resource group...")
+		log.Println("recreating E2E test resource group...")
 
 		needNewCluster = true
 		_, err := cloud.resourceGroupClient.CreateOrUpdate(
@@ -68,7 +68,7 @@ func setupCluster(ctx context.Context, t *testing.T, cloud *azureClient, locatio
 
 	// A new cluster is created if the test RG does not exist, the cluster itself does not exist, or if the cluster is in an unusable state
 	if needNewCluster {
-		t.Log("recreating E2E test cluster... ")
+		log.Println("recreating E2E test cluster... ")
 
 		pollerResp, err := cloud.aksClient.BeginCreateOrUpdate(
 			ctx,
