+++ e2e/vmss.go
@@ -1,4 +1,4 @@
-package e2e_test
+package main
 
 import (
 	"context"
